% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeReturns}
\alias{computeReturns}
\title{Compute logarithmic returns for a given window width.}
\usage{
computeReturns(daily_price_data, window_width = 1L)
}
\arguments{
\item{daily_price_data}{(data frame) daily price data with `date`, `symbol`, `close` columns}

\item{window_width}{(integer) number of periods width of aggregation window}
}
\description{
Compute logarithmic returns for a given window width.
}
