% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{interCurrencyDifferences}
\alias{interCurrencyDifferences}
\title{Assess momentum pattern differences between cryptocurrencies}
\usage{
interCurrencyDifferences(
  computed_returns = computeReturns(daily_price_data = importAllYahooFinanceFiles()),
  min_observations_num = 180L
)
}
\description{
Assess momentum pattern differences between cryptocurrencies
}
