% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{interpretRegression}
\alias{interpretRegression}
\title{Interpret linear regression results and output a formatted data frame}
\usage{
interpretRegression(lm_result)
}
\arguments{
\item{lm_result}{(lm) result of running `lm`}
}
\description{
Interpret linear regression results and output a formatted data frame
}
