% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{momentumPersistence}
\alias{momentumPersistence}
\title{Test of whether momentum patterns persist beyond a training period}
\usage{
momentumPersistence(
  dat,
  estimation_months = 6L,
  validation_months = 3L,
  window_width = 1L
)
}
\description{
Test of whether momentum patterns persist beyond a training period
}
