% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{priceMovePlot}
\alias{priceMovePlot}
\title{Generate price plot}
\usage{
priceMovePlot(
  data,
  smb_list,
  start_date = as.Date("1900-01-01"),
  end_date = as.Date("2900-01-01"),
  type = c("open", "high", "low", "close"),
  brk = "2 month"
)
}
\arguments{
\item{data}{(data frame) daily price data with `date`, `symbol`, `open`, `high`, `low`, `close` columns}

\item{smb_list}{(character vector) list of symbols for which generate plot}

\item{start_date}{(character) start date in format "%Y-%m-%d" for data in plot}

\item{end_date}{(character) end date in format "%Y-%m-%d" for data in plot}

\item{type}{(character) type of price for plot from: c("open","high","low","close")}

\item{brk}{(character) how often show labels on x-axis}
}
\description{
Generate price plot
}
