% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{returnNtilesNoLookAhead}
\alias{returnNtilesNoLookAhead}
\title{Ntile analysis with no lookahead:w}
\usage{
returnNtilesNoLookAhead(
  returns_data,
  training_date_max = "2015-01-01",
  ntile_n = 5L,
  leads = 1:12L
)
}
\description{
Ntile analysis with no lookahead:w
}
