% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{runAllDailyRegressions}
\alias{runAllDailyRegressions}
\title{Run all daily regressions per symbol at indicated leads}
\usage{
runAllDailyRegressions(returns_data, leads = 1:12L)
}
\description{
Run all daily regressions per symbol at indicated leads
}
