% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{runRegressionLeads}
\alias{runRegressionLeads}
\title{Run linear regressions for different lead values}
\usage{
runRegressionLeads(returns_data, leads = 1:12L)
}
\arguments{
\item{leads}{(integer) lead values for which to run analysis}

\item{reaturns_data}{(data frame) with periodic returns data}
}
\description{
Run linear regressions for different lead values
}
