% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summariseMomentumPrevalence}
\alias{summariseMomentumPrevalence}
\title{Summarise results of momentum prevalence analysis}
\usage{
summariseMomentumPrevalence(prevalence_results)
}
\description{
Summarise results of momentum prevalence analysis
}
